pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signaldeck-plugin-main"
version = "0.1.0"
description = "Main plugin for SignalDeck"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "signaldeck-sdk>=0.1.0",
  "signaldeck-ui>=0.1.0",
  "flask>=3.0",
  "paho-mqtt>=1.6.1",
  "pymodbus>=3.8.6",
  "jsonpath-ng>=1.6.1",
  "requests>=2.32.3"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["signaldeck_plugin_main*"]

[tool.setuptools.package-data]
signaldeck_plugin_main = [
  "templates/**/*"
]
